Fix the position saving of the separtor between the tree and the unknown index list
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/Thaw@11598 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/thaw/plugins/index/IndexBrowserPanel.java
@@ -136,7 +136,7 @@ public class IndexBrowserPanel implements javax.swing.event.TreeSelectionListene
 		splitLocation = leftSplit.getDividerLocation();
 
 		config.setValue("indexTreeUnknownListSplitLocation",
-				Double.toString(splitLocation));
+				Integer.toString(splitLocation));
 
 		tables.saveState();
 	}
